Integrated PluginEditor from GUI branch
This GUI has two buttons Chorus and Reverb that are two processors that will be used to populate the editor window. As of right now, when a user clicks Chorus a display 'Chorus Selected' appears in the console as well as a function call that returns and displays a NodeID of the newly created processor node. DO NOT OVERUSE, this is purely for testing purposes only.